+++ packages/account-partner/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-partner
 
+## 3.0.0-next.25
+
+### Patch Changes
+
+- Adding all chains as supported by the connector
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.25
+
 ## 3.0.0-next.24
 
 ### Patch Changes