+++ packages/account-react-native/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sophon-labs/account-react-native
 
+## 3.0.0-next.25
+
+### Patch Changes
+
+- Adding all chains as supported by the connector
+- Updated dependencies
+  - @sophon-labs/account-message-bridge@3.0.0-next.25
+  - @sophon-labs/account-communicator@3.0.0-next.25
+  - @sophon-labs/account-connector@3.0.0-next.25
+  - @sophon-labs/account-provider@3.0.0-next.25
+  - @sophon-labs/account-partner@3.0.0-next.25
+  - @sophon-labs/account-core@3.0.0-next.25
+
 ## 3.0.0-next.24
 
 ### Patch Changes